+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 # PayPal iOS SDK Release Notes
 
 ## unreleased
+* PayPalWebPayments
+  * Add `vault(url:)` method to `PayPalWebCheckoutClient`
+  * Add `PayPalVaultResult` type to return vault result
+  * Add `PayPalVaultDelegate` to handle results from vault flow
+  * Add `PayPalWebCheckoutClientError.paypalVaultResponseError` for missing or invalid response from vaulting
 * PaymentButtons
   * Add `custom` case for `PaymentButtonEdges`